+++ app/src/test/java/co/kr/woowahan_banchan/data/repository/fakedatasource/FakeOrderDataSource.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_banchan.data.repository
+package co.kr.woowahan_banchan.data.repository.fakedatasource
 
 import co.kr.woowahan_banchan.data.datasource.local.order.OrderDataSource
 import co.kr.woowahan_banchan.data.model.local.OrderDto
